Prints an overview with summary of model class 'hdlmm'

Method for printing an overview with summary of model class 'hdlmm'

Usage

# S3 method for summary.hdlmm
print(x, digits = 3, cw.only = TRUE, ...)

Arguments

x

an object of type 'summary.hdlmm', result of call to summary.hdlmm()

digits

integer number of digits to round

cw.only

print only results for exposures with critical windows

...

additional parameters

Value

output of hdlmm fit in R console

Details

print.summary.hdlmm

Examples

D <- sim.hdlmm(sim = "D", n = 1000)
fit <- dlmtree(y ~ ., 
               data = D$dat,
               exposure.data = D$exposures,
               dlm.type = "linear",
               family = "gaussian",
               mixture = TRUE,
               het = TRUE)
#> Preparing data...
#> 
#> Running HDLMM:
#> Burn-in % complete 
#> [0--------25--------50--------75--------100]
#>  ''''''''''''''''''''''''''''''''''''''''''
#> MCMC iterations (est time: 10 seconds)
#> [0--------25--------50--------75--------100]
#>  ''''''''''''''''''''''''''''''''''''''''''
#> Compiling results...
fit_sum <- summary(fit)
print(fit_sum)
#> ---
#> HDLMM summary
#> 
#> Model run info:
#> - y ~ . 
#> - family: gaussian 
#> - 20 trees
#> - 1000 burn-in iterations
#> - 2000 post-burn iterations
#> - 2 thinning factor
#> - 3 exposures measured at 37 time points
#> - 3 two-way interactions (no-self interactions)
#> - 0.5 modifier sparsity prior
#> - 1 exposure sparsity prior
#> - 0.95 confidence level
#> 
#> Fixed effects:
#>                Mean  Lower  Upper
#>  (Intercept)  0.100 -0.298  0.484
#> *mod_num     -0.209 -0.427 -0.052
#> *mod_bin      0.684  0.460  0.929
#> *mod_scale    1.214  0.932  1.487
#> *c1           0.181  0.097  0.263
#> *c2           0.611  0.522  0.690
#> *c3           0.307  0.226  0.387
#> *c4           0.247  0.163  0.343
#>  c5          -0.010 -0.094  0.060
#> *b1          -1.003 -1.168 -0.845
#> *b2          -0.241 -0.413 -0.057
#>  b3          -0.058 -0.231  0.106
#> *b4           1.486  1.289  1.665
#> *b5           0.204  0.048  0.347
#> ---
#> * = CI does not contain zero
#> 
#> Modifiers:
#>             PIP
#> mod_num   0.906
#> mod_bin   0.844
#> mod_scale 0.628
#> c1        0.643
#> c2        0.688
#> c3        0.652
#> c4        0.546
#> c5        0.712
#> b1        0.593
#> b2        0.673
#> b3        0.658
#> b4        0.582
#> b5        0.556
#> ---
#> PIP = Posterior inclusion probability
#> 
#> ---
#> To obtain exposure effect estimates, use the 'shiny(fit)' function.
#>