Prints an overview with summary of model class 'monotone'

Method for printing an overview with summary of model class 'monotone'

Usage

# S3 method for summary.monotone
print(x, digits = 3, ...)

Arguments

x

an object of type 'summary.monotone', result of call to summary.monotone()

digits

integer number of digits to round

...

additional parameters

Value

output in R console

Details

print.summary.monotone

Examples

D <- sim.tdlnm(sim = "A", error.to.signal = 1)
fit <- dlmtree(formula = y ~ .,
               data = D$dat,
               exposure.data = D$exposures,
               dlm.type = "monotone",
               family = "gaussian")
#> Preparing data...
#> 
#> Running monotone-TDLNM:
#> Burn-in % complete 
#> [0--------25--------50--------75--------100]
#>  ''''''''''''''''''''''''''''''''''''''''''
#> MCMC iterations (est time: 4 seconds)
#> [0--------25--------50--------75--------100]
#>  ''''''''''''''''''''''''''''''''''''''''''
#> Compiling results...
fit_sum <- summary(fit)
#> Centered DLNM at exposure value 0 
print(fit_sum)
#> ---
#> monotone-TDLNM summary
#> 
#> Model run info:
#> - y ~ . 
#> - sample size: 926 
#> - family: gaussian 
#> - 20 trees
#> - 1000 burn-in iterations
#> - 2000 post-burn iterations
#> - 2 thinning factor
#> - 0.95 confidence level
#> 
#> Fixed effect coefficients:
#>                Mean  Lower  Upper
#> *(Intercept) -3.210 -3.769 -2.767
#> c1            0.122 -0.013  0.256
#> *c2           0.433  0.280  0.584
#> *c3          -0.908 -1.045 -0.753
#> *c4          -1.874 -2.012 -1.722
#> *c5          -0.442 -0.596 -0.296
#> *b1          -1.913 -2.205 -1.609
#> b2            0.238 -0.034  0.513
#> b3           -0.037 -0.332  0.253
#> b4            0.152 -0.172  0.441
#> b5           -0.105 -0.370  0.164
#> ---
#> * = CI does not contain zero
#> 
#> DLNM effect:
#> range = [0.001, 0.006]
#> signal-to-noise = 0
#> critical windows: No critical windows 
#> 
#> residual standard errors: 0.299