Prints an overview with summary of model class 'tdlm'

Method for printing an overview with summary of model class 'tdlm'

Usage

# S3 method for summary.tdlm
print(x, digits = 3, ...)

Arguments

x

an object of type 'summary.tdlnm', result of call to summary.tdlnm()

digits

integer number of digits to round

...

additional parameters

Value

output of tdlm fit in R console

Details

print.summary.tdlm

Examples

D <- sim.tdlmm(sim = "A", mean.p = 0.5, n = 1000)
fit <- dlmtree(y ~ .,
               data = D$dat,
               exposure.data = D$exposures[[1]],
               dlm.type = "linear",
               family = "logit",
               binomial.size = 1)
#> Preparing data...
#> 
#> Running TDLM:
#> Burn-in % complete 
#> [0--------25--------50--------75--------100]
#>  ''''''''''''''''''''''''''''''''''''''''''
#> MCMC iterations (est time: 2 seconds)
#> [0--------25--------50--------75--------100]
#>  ''''''''''''''''''''''''''''''''''''''''''
#> Compiling results...
fit_sum <- summary(fit)
print(fit_sum)
#> ---
#> TDLM summary
#> 
#> Model run info:
#> - y ~ . 
#> - family: logit 
#> - 20 trees
#> - 1000 burn-in iterations
#> - 2000 post-burn iterations
#> - 2 thinning factor
#> - 0.95 confidence level
#> 
#> Fixed effect coefficients:
#>                Mean  Lower  Upper
#> *(Intercept)  0.357  0.008  0.688
#> c1           -0.097 -0.232  0.043
#> *c2           0.257  0.128  0.383
#> *c3          -0.140 -0.274 -0.012
#> c4            0.037 -0.102  0.173
#> c5            0.087 -0.044  0.224
#> b1           -0.132 -0.391  0.135
#> b2           -0.173 -0.440  0.090
#> b3           -0.084 -0.357  0.180
#> *b4          -0.303 -0.561 -0.043
#> b5           -0.073 -0.329  0.170
#> ---
#> * = CI does not contain zero
#> 
#> DLM effect:
#> range = [-0.025, 0.133]
#> signal-to-noise = 0.3
#> critical windows: 23-29 
#>              Mean  Lower Upper
#> Period 1   -0.008 -0.066 0.069
#> Period 2   -0.011 -0.059 0.048
#> Period 3   -0.015 -0.057 0.034
#> Period 4   -0.019 -0.061 0.022
#> Period 5   -0.025 -0.074 0.011
#> Period 6   -0.021 -0.067 0.017
#> Period 7   -0.016 -0.055 0.035
#> Period 8   -0.018 -0.054 0.023
#> Period 9   -0.020 -0.055 0.015
#> Period 10  -0.018 -0.056 0.019
#> Period 11  -0.019 -0.058 0.017
#> Period 12  -0.022 -0.079 0.017
#> Period 13  -0.017 -0.066 0.024
#> Period 14  -0.008 -0.044 0.041
#> Period 15  -0.009 -0.044 0.033
#> Period 16  -0.009 -0.050 0.031
#> Period 17  -0.007 -0.053 0.034
#> Period 18   0.000 -0.050 0.047
#> Period 19   0.012 -0.036 0.080
#> Period 20   0.020 -0.036 0.088
#> Period 21   0.038 -0.028 0.117
#> Period 22   0.061 -0.021 0.142
#> *Period 23  0.115  0.022 0.196
#> *Period 24  0.133  0.078 0.224
#> *Period 25  0.131  0.079 0.208
#> *Period 26  0.115  0.058 0.165
#> *Period 27  0.110  0.050 0.165
#> *Period 28  0.106  0.040 0.169
#> *Period 29  0.097  0.017 0.180
#> Period 30   0.038 -0.073 0.114
#> Period 31   0.028 -0.030 0.083
#> Period 32   0.025 -0.025 0.078
#> Period 33   0.022 -0.026 0.069
#> Period 34   0.021 -0.026 0.071
#> Period 35   0.012 -0.049 0.063
#> Period 36   0.008 -0.056 0.062
#> Period 37   0.008 -0.059 0.068
#> ---
#> * = CI does not contain zero