Prints an overview with summary of model class 'tdlmm'

Method for printing an overview with summary of model class 'tdlmm'

Usage

# S3 method for summary.tdlmm
print(x, digits = 3, cw.only = TRUE, ...)

Arguments

x

an object of type 'summary.tdlmm', result of call to summary.tdlmm()

digits

integer number of digits to round

cw.only

print only results for exposures with critical windows

...

additional parameters

Value

output of tdlmm fit in R console

Details

print.summary.tdlmm

Examples

D <- sim.tdlmm(sim = "B", error = 25, n = 1000)
fit <- dlmtree(y ~ .,
               data = D$dat, exposure.data = D$exposures,
               mixture.interactions = "noself", 
               dlm.type = "linear", family = "gaussian",
               mixture = TRUE)
#> Preparing data...
#> 
#> Running TDLMM:
#> Burn-in % complete 
#> [0--------25--------50--------75--------100]
#>  ''''''''''''''''''''''''''''''''''''''''''
#> MCMC iterations (est time: 6 seconds)
#> [0--------25--------50--------75--------100]
#>  ''''''''''''''''''''''''''''''''''''''''''
#> Compiling results...
fit_sum <- summary(fit)
#> Reconstructing main effects...
#> Reconstructing interaction effects...
#> 0%...25%...50%...75%...100%
#> Calculating marginal effects...
#> Calculating fixed effects...
print(fit_sum)
#> ---
#> TDLMM summary
#> 
#> Model run info:
#> - y ~ . 
#> - family: gaussian 
#> - 20 trees (alpha = 0.95, beta = 2)
#> - 1000 burn-in iterations
#> - 2000 post-burn iterations
#> - 2 thinning factor
#> - 5 exposures measured at 37 time points
#> - 10 two-way interactions (no-self interactions)
#> - 1 kappa sparsity prior
#> - 0.95 confidence level
#> 
#> Fixed effects:
#>                Mean  Lower  Upper
#> *(Intercept)  6.680  3.604  9.726
#> *c1          -0.894 -1.210 -0.575
#> *c2          -1.683 -1.998 -1.385
#> *c3          -1.549 -1.884 -1.215
#>  c4          -0.138 -0.458  0.180
#>  c5          -0.259 -0.575  0.052
#> *b1           1.452  0.801  2.100
#>  b2           0.023 -0.684  0.696
#>  b3           0.538 -0.066  1.162
#> *b4          -0.782 -1.346 -0.134
#>  b5           0.265 -0.392  0.903
#> ---
#> * = CI does not contain zero
#> 
#> --
#> Exposure effects: critical windows
#> * = Exposure selected by Bayes Factor
#> (x.xx) = Relative effect size
#> 
#>  *e1 (0.52): 14-19
#>  *e2 (0.5): 
#>  *e3 (0.19): 
#>  *e4 (0.58): 
#>  *e5 (0.71): 
#> --
#> Interaction effects: critical windows
#> 
#>  - No critical windows
#> ---