Determines split points for continuous modifiers

Method for determining split points for continuous modifiers

Usage

splitpoints(object, var, round = NULL)

Arguments

object

An object of class 'hdlm', 'hdlmm'

var

The name of a continuous variable for which the split points will be reported

round

The number of decimal places to round the variable (var) to. No rounding occurs if round=NULL (default) For positive integer values of round, the variable will be rounded and split points will be reported at the resulting level

Value

A data frame with split points and the probability that a split point was >= that split point value

Details

splitpoints

Examples

# \donttest{
# Split points with HDLM 
D <- sim.hdlmm(sim = "B", n = 1000)
fit <- dlmtree(y ~ ., 
               data = D$dat,
               exposure.data = D$exposures,
               dlm.type = "linear",
               family = "gaussian",
               het = TRUE)
#> Preparing data...
#> 
#> Running shared HDLM:
#> Burn-in % complete 
#> [0--------25--------50--------75--------100]
#>  ''''''''''''''''''''''''''''''''''''''''''
#> MCMC iterations (est time: 6 seconds)
#> [0--------25--------50--------75--------100]
#>  ''''''''''''''''''''''''''''''''''''''''''
#> Compiling results...
splitpoints(fit, var = "mod_num", round = 2)
#>    location  proportion
#> 1     -1.65 0.039206850
#> 2     -1.30 0.034700315
#> 3     -1.01 0.030644434
#> 4     -0.84 0.038756196
#> 5     -0.66 0.044614691
#> 6     -0.52 0.047769265
#> 7     -0.39 0.037854890
#> 8     -0.27 0.052275800
#> 9     -0.16 0.050022533
#> 10    -0.03 0.132942767
#> 11     0.10 0.094637224
#> 12     0.21 0.048670572
#> 13     0.32 0.059486255
#> 14     0.48 0.063091483
#> 15     0.62 0.066696710
#> 16     0.74 0.045515998
#> 17     0.89 0.027489860
#> 18     1.12 0.039206850
#> 19     1.35 0.043713384
#> 20     1.76 0.002703921
splitpoints(fit, var = "mod_scale", round = 2)
#>    location  proportion
#> 1      0.05 0.059292649
#> 2      0.10 0.046463245
#> 3      0.14 0.047156727
#> 4      0.20 0.051664355
#> 5      0.26 0.079403606
#> 6      0.31 0.074202497
#> 7      0.36 0.032246879
#> 8      0.41 0.040915395
#> 9      0.46 0.030859917
#> 10     0.51 0.032593620
#> 11     0.55 0.053398058
#> 12     0.60 0.047850208
#> 13     0.64 0.073509015
#> 14     0.68 0.044036061
#> 15     0.74 0.041955617
#> 16     0.79 0.051664355
#> 17     0.84 0.081484050
#> 18     0.88 0.045769764
#> 19     0.92 0.058945908
#> 20     0.96 0.006588072
# }